% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{index}
\alias{index}
\title{Index Geonames data.}
\usage{
index(geonames_file)
}
\arguments{
\item{geonames_file}{Path to the Geonames raw CSV.}
}
\description{
This function indexes the raw data (in CSV format) extracted from 
Geonames into a Lucene++ index.
}
\note{
To learn more about the Geonames data, please use the following link:
      https://download.geonames.org/export/dump/readme.txt

To download the data, please access the Geonames website:
      https://download.geonames.org/export/dump/
}
\references{
GeoNames. GeoNames. http://geonames.org/. Retr. February 11, 2024
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
