% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{search}
\alias{search}
\title{Search Geonames data.}
\usage{
search(text, max_items = 100)
}
\arguments{
\item{text}{Query text}

\item{max_items}{Max number of files to return as result.}
}
\value{
Search result items.
}
\description{
This function searches content across all Geonames data indexed locally.
}
\note{
To use this function, you must have a local Geonames index. To learn
      more, please check [geonamesr::index].

To learn more about the query syntax, please check the Lucene core 
      documentation (https://lucene.apache.org/core/9_9_2/index.html).
}
\references{
GeoNames. GeoNames. http://geonames.org/. Retr. February 11, 2024
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
